Fix leak in `VectorArray` -> `Vec<Array>` conversion
All other `MaybeUninitFoo::try_into_guarded()` implementations
just hand ownership of raw resources to `Foo`, so
`MaybeUninitFoo` doesn't need to free the resource itself.

`MaybeUninitVectorArray` -> `Vec<Array>` is different:
it returns the copies of individual arrays without giving up
the ownership of the original `VectorArray`,
so it remains responsible for freeing it afterwards.